requirements.txt
git+https://github.com/huggingface/transformers.git
datasets
sentencepiece
pytorch-lightning
wandb
